lib: refacto rename + use TMP DIR for pygen synthdefs + some cleaning
