+++ renardo_lib/renardo_lib/Extensions/SonicPiSynthDefImporter/generate_sonic_pi.py
@@ -1,4 +1,4 @@
-from renardo_lib.Extensions.SonicPi import GenerateSonicPiData
+from renardo_lib.Extensions.SonicPiSynthDefImporter import GenerateSonicPiData
 
 if __name__ == '__main__':
     import sys
