--- renardo_lib/renardo_lib/SynthDefManagement/Effects/__init__.py
@@ -1,1 +1,0 @@
-from renardo_lib.SynthDefManagement.Effects.Util import *
