+++ renardo_lib/renardo_lib/SynthDefManagement/SCLangExperimentalPythonBindings/NewPygenEffectClassUnused.py
@@ -3,8 +3,6 @@ import inspect
 from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings import cls, Vibrato, RHPF
 from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings.core import instance
 
-# To be moved to SynthDefManagement later
-
 EnvGen = cls("EnvGen")
 Env = cls("Env")
