+++ renardo_lib/renardo_lib/SynthDefManagement/__init__.py
@@ -2,4 +2,5 @@ from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings import *
 from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings.PygenSynthDef import DefaultPygenSynthDef
 from renardo_lib.SynthDefManagement.SynthDefProxy import SynthDefProxy
 from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings.SamplePygenSynthDefs import SamplePygenSynthDef
+from renardo_lib.SynthDefManagement.SCLangExperimentalPythonBindings.PygenEffectSynthDefs import *
