+++ renardo_lib/renardo_lib/runtime/synthdefs_initialisation/__init__.py
@@ -1,5 +1,7 @@
-from .special_synthdefs import *
-from .python_defined_synthdefs import *
+from renardo_lib.SynthDefManagement.SimpleSynthDefs import LiveSynthDef
+from .python_defined_effect_synthdefs import *
+#from .special_synthdefs import *
+#from .python_defined_synthdefs import *
 from .sclang_file_synthdefs import *
