+++ renardo_lib/renardo_lib/runtime/synthdefs_initialisation/crashFX.py
@@ -1,71 +1,72 @@
 # Legato slide
-fx = FxList.new("leg", "leg", {"leg":0, "sus":1 }, order = 0)
+
+fx = effect_manager.new("leg", "leg", {"leg":0, "sus":1 }, order = 0)
 fx.add("osc = osc * XLine.ar(Rand(0.5,1.5)*leg,1,0.05*sus)")
 fx.save()
 
 # Lpf slide
-fx = FxList.new('spf','SLPF', {'spf': 0, 'spr': 1, 'spfslide':1, 'spfend':15000}, order=2)
+fx = effect_manager.new('spf','SLPF', {'spf': 0, 'spr': 1, 'spfslide':1, 'spfend':15000}, order=2)
 fx.add_var("spfenv")
 fx.add('spfenv = EnvGen.ar(Env.new([spf, spfend], [spfslide]))')
 fx.add('osc = RLPF.ar(osc, spfenv, spr)')
 fx.save()
 
 # MoogLPF
-fx = FxList.new('mpf','MoogFF', {'mpf': 0, 'mpr': 0}, order=2)
+fx = effect_manager.new('mpf','MoogFF', {'mpf': 0, 'mpr': 0}, order=2)
 fx.doc("MoogFF filter")
 fx.add('osc = MoogFF.ar(osc, mpf, mpr,0,1)')
 fx.save()
 
 # DFM1 LPF
-fx = FxList.new('dfm','DFM1', {'dfm': 1000, 'dfmr': 0.1, 'dfmd': 1}, order=2)
+fx = effect_manager.new('dfm','DFM1', {'dfm': 1000, 'dfmr': 0.1, 'dfmd': 1}, order=2)
 fx.doc("DFM1 filter")
 fx.add('osc = DFM1.ar(osc, dfm, dfmr, dfmd,0.0)')
 fx.save()
 
 if SC3_PLUGINS:
     #Dist mod
-    fx = FxList.new('disto', 'disto_mod', {'disto': 0, 'smooth': 0.3, 'distomix': 1}, order=1)
+    fx = effect_manager.new('disto', 'disto_mod', {'disto': 0, 'smooth': 0.3, 'distomix': 1}, order=1)
     fx.add("osc = LinXFade2.ar(CrossoverDistortion.ar(osc, amp:0.5*disto, smooth:smooth),  osc, 1-distomix)")
     fx.save()
 
 #New Chop, with wave select :
 #chopwave = (0: Pulse, 1: Tri, 2: Saw, 3: Sin, 4: Parabolic )
 # and chopi = oscillator phase
-fx = FxList.new('chop', 'chop', {'chop': 0, 'sus': 1, 'chopmix': 1, 'chopwave': 0, 'chopi': 0}, order=2)
+fx = effect_manager.new('chop', 'chop', {'chop': 0, 'sus': 1, 'chopmix': 1, 'chopwave': 0, 'chopi': 0}, order=2)
 fx.add("osc = LinXFade2.ar(osc * SelectX.kr(chopwave, [LFPulse.kr(chop / sus, iphase:chopi, add: 0.01), LFTri.kr(chop / sus, iphase:chopi, add: 0.01), LFSaw.kr(chop / sus, iphase:chopi, add: 0.01), FSinOsc.kr(chop / sus, iphase:chopi, add: 0.01), LFPar.kr(chop / sus, iphase:chopi, add: 0.01)]), osc, 1-chopmix)")
 fx.save()
 
 
-fx = FxList.new('tremolo', 'tremolo', {'tremolo': 0, 'beat_dur': 1, 'tremolomix': 1}, order=2)
+fx = effect_manager.new('tremolo', 'tremolo', {'tremolo': 0, 'beat_dur': 1, 'tremolomix': 1}, order=2)
 fx.add("osc = LinXFade2.ar(osc * SinOsc.ar( tremolo / beat_dur, mul:0.5, add:0.5), osc, 1-tremolomix)")
 fx.save()
 
-fx = FxList.new('echo', 'combDelay', {'echo': 0, 'echomix' : 1, 'beat_dur': 1, 'echotime': 1}, order=2)
+fx = effect_manager.new('echo', 'combDelay', {'echo': 0, 'echomix' : 1, 'beat_dur': 1, 'echotime': 1}, order=2)
 fx.add('osc = LinXFade2.ar(osc + CombL.ar(osc, delaytime: echo * beat_dur, maxdelaytime: 2 * beat_dur, decaytime: echotime * beat_dur), osc, 1-echomix)')
 fx.save()
 
-fx = FxList.new('flanger', 'flanger', {'flanger': 0, 'fdecay': 0, 'flangermix':1}, order=2)
+fx = effect_manager.new('flanger', 'flanger', {'flanger': 0, 'fdecay': 0, 'flangermix':1}, order=2)
 fx.add("osc = LinXFade2.ar(CombC.ar(osc, 0.01, SinOsc.ar(flanger, 0, (0.01 * 0.5) - 0.001, (0.01 * 0.5) + 0.001), fdecay, 1),  osc, 1-flangermix)")
 fx.save()
 
-fx = FxList.new("formant", "formantFilter", {"formant": 0, 'formantmix': 1}, order=2)
+fx = effect_manager.new("formant", "formantFilter", {"formant": 0, 'formantmix': 1}, order=2)
 fx.add("formant = (formant % 8) + 1")
 fx.add("osc = LinXFade2.ar(Formlet.ar(osc, formant * 200, ((formant % 5 + 1)) / 1000, (formant * 1.5) / 600).tanh, osc, 1-formantmix)")
 fx.save()
 
-fx = FxList.new("shape", "wavesShapeDistortion", {"shape":0, "shapemix":1}, order=2)
+fx = effect_manager.new("shape", "wavesShapeDistortion", {"shape":0, "shapemix":1}, order=2)
 fx.add("osc = LinXFade2.ar((osc * (shape * 50)).fold2(1).distort / 5, osc, 1-shapemix)")
 fx.save()
 
-fx = FxList.new("drive", "overdriveDistortion", {"drive":0, "drivemix":1}, order=2)
+fx = effect_manager.new("drive", "overdriveDistortion", {"drive":0, "drivemix":1}, order=2)
 fx.add("osc = LinXFade2.ar((osc * (drive * 50)).clip(0,0.2).fold2(2), osc, 1-drivemix)")
 fx.save()
 
-fx = FxList.new("tanh", "tanhDisto", {"tanh":0}, order=2)
+fx = effect_manager.new("tanh", "tanhDisto", {"tanh":0}, order=2)
 fx.add("osc = osc + (osc*tanh).tanh.sqrt()")
 fx.save()
 
-fx = FxList.new("fdist", "fdist", {"fdist":0, "fdistfreq":1600}, order=2)
+fx = effect_manager.new("fdist", "fdist", {"fdist":0, "fdistfreq":1600}, order=2)
 fx.add("osc = LPF.ar(osc, fdistfreq)")
 fx.add("osc = (osc * 1.1 * fdist).tanh")
 fx.add("osc = LPF.ar(osc, fdistfreq)")
@@ -77,7 +78,7 @@ fx.add("osc = (osc * 2 * fdist).tanh")
 fx.add("osc = osc*0.2")
 fx.save()
 
-fx = FxList.new("fdistc", "fdistc", {"fdistc":0, "fdistcfreq1":1600,"fdistcfreq2":1600,"fdistcfreq3":1600,"fdistcfreq4":1600, "fdistcm1": 1.1, "fdistcm2": 1.1, "fdistcm3": 1.4, "fdistcm4": 2, "fdistcq1": 1, "fdistcq2": 1, "fdistcq3":1, "fdistcq4":1}, order=2)
+fx = effect_manager.new("fdistc", "fdistc", {"fdistc":0, "fdistcfreq1":1600,"fdistcfreq2":1600,"fdistcfreq3":1600,"fdistcfreq4":1600, "fdistcm1": 1.1, "fdistcm2": 1.1, "fdistcm3": 1.4, "fdistcm4": 2, "fdistcq1": 1, "fdistcq2": 1, "fdistcq3":1, "fdistcq4":1}, order=2)
 fx.add("osc = RLPF.ar(osc, fdistcfreq1, fdistcq1)")
 fx.add("osc = (osc * fdistcm1 * fdistc).tanh")
 fx.add("osc = RLPF.ar(osc, fdistcfreq2, fdistcq2)")
@@ -89,7 +90,7 @@ fx.add("osc = (osc * fdistcm4 * fdistc).tanh")
 fx.save()
 
 #based on Derek Kwan chorus
-fx = FxList.new("chorus", "chorus", {"chorus":0, "chorusrate":0.5}, order=2)
+fx = effect_manager.new("chorus", "chorus", {"chorus":0, "chorusrate":0.5}, order=2)
 fx.add_var("lfos")
 fx.add_var("numDelays = 4")
 fx.add_var("chrate")
@@ -105,14 +106,14 @@ fx.add("osc = Mix(osc)")
 fx.save()
 
 # dub delay based on «Dub Echo» by Bjorn Westergard [sccode https://sccode.org/1-h]
-fx = FxList.new('dubd', 'dubdelay', {'dubd': 0, 'dublen': 0.1, 'dubwidth': 0.12, 'dubfeed': 0.8}, order=2)
+fx = effect_manager.new('dubd', 'dubdelay', {'dubd': 0, 'dublen': 0.1, 'dubwidth': 0.12, 'dubfeed': 0.8}, order=2)
 fx.add_var("dry")
 fx.add("dry = osc")
 fx.add("osc = osc + Fb({ |feedback| var left, right; var magic = LeakDC.ar(feedback*dubfeed + osc); magic = HPF.ar(magic, 400); magic = LPF.ar(magic, 5000); magic = magic.tanh; #left, right = magic; magic = [DelayC.ar(left, 1, LFNoise2.ar(12).range(0,dubwidth)), DelayC.ar(right, 1, LFNoise2.ar(12).range(dubwidth,0))].reverse;	},dublen)")
 fx.add("osc = SelectX.ar(dubd, [dry, osc])")
 fx.save()
 
-fx = FxList.new('octafuz', 'octafuz', {'octafuz': 0, 'octamix': 1}, order=2)
+fx = effect_manager.new('octafuz', 'octafuz', {'octafuz': 0, 'octamix': 1}, order=2)
 fx.add_var("dis")
 fx.add_var("osc_base")
 fx.add("osc_base = osc")
@@ -123,7 +124,7 @@ fx.add("osc = (osc * 1/16)!2")
 fx.add("osc = LinXFade2.ar(osc_base, osc, octamix)")
 fx.save()
 
-fx = FxList.new('tek', 'tek', {'tek': 0, 'tekr':500, 'tekd':8}, order=2)
+fx = effect_manager.new('tek', 'tek', {'tek': 0, 'tekr':500, 'tekd':8}, order=2)
 fx.add_var("osc_low")
 fx.add_var("osc_med")
 fx.add_var("osc_high")
@@ -145,7 +146,7 @@ fx.save()
 
 
 ### TIDAL FX ####
-fx = FxList.new("krush", "dirt_krush", {"krush":0, "kutoff":15000}, order=2)
+fx = effect_manager.new("krush", "dirt_krush", {"krush":0, "kutoff":15000}, order=2)
 fx.add_var("signal")
 fx.add_var("freq")
 fx.add("freq = Select.kr(kutoff > 0, [DC.kr(4000), kutoff])")
@@ -154,15 +155,15 @@ fx.add("signal = RLPF.ar(signal, clip(freq, 20, 10000), 1)")
 fx.add("osc = SelectX.ar(krush * 2.0, [osc, signal])")
 fx.save()
 
-fx = FxList.new("drop", "waveloss", {"drop":0, "dropof": 100}, order=2)
+fx = effect_manager.new("drop", "waveloss", {"drop":0, "dropof": 100}, order=2)
 fx.add("osc = WaveLoss.ar(osc, drop, outof: dropof, mode: 2)")
 fx.save()
 
-fx = FxList.new("squiz", "squiz", {"squiz":0}, order=2)
+fx = effect_manager.new("squiz", "squiz", {"squiz":0}, order=2)
 fx.add("osc = Squiz.ar(osc, squiz)")
 fx.save()
 
-fx = FxList.new("triode", "triode", {"triode":0}, order=2)
+fx = effect_manager.new("triode", "triode", {"triode":0}, order=2)
 fx.add_var("sc")
 fx.add("sc = triode * 10 + 1e-3")
 fx.add("osc = (osc * (osc > 0)) + (tanh(osc * sc) / sc * (osc < 0))")
@@ -170,7 +171,7 @@ fx.add("osc = LeakDC.ar(osc)*1.2")
 fx.save()
 
 #### Need Tweak ##############
-fx = FxList.new("octer", "octer", {"octer":0, "octersub": 0, "octersubsub": 0}, order=1)
+fx = effect_manager.new("octer", "octer", {"octer":0, "octersub": 0, "octersubsub": 0}, order=1)
 fx.add_var("oct1")
 fx.add_var("oct2")
 fx.add_var("oct3")
@@ -183,7 +184,7 @@ fx.add("osc = SelectX.ar(octer, [osc, octer*oct1, DC.ar(0)])")
 fx.add("osc = osc + (octersub * oct2 * sub) + (octersubsub * oct3 * sub)")
 fx.save()
 
-fx = FxList.new("feed", "feeddelay", {"feed":0.7, "feedfreq": 50}, order=2)
+fx = effect_manager.new("feed", "feeddelay", {"feed":0.7, "feedfreq": 50}, order=2)
 fx.add_var("out")
 fx.add("out = osc + Fb({\
 		arg feedback;\
@@ -191,37 +192,37 @@ fx.add("out = osc + Fb({\
 	},0.5,0.125)")
 fx.save()
 
-fx = FxList.new("sample_atk", "sample_atk", {"sample_atk":0, "sample_sus":1}, order=2)
+fx = effect_manager.new("sample_atk", "sample_atk", {"sample_atk":0, "sample_sus":1}, order=2)
 fx.add_var("env")
 fx.add("env = EnvGen.ar(Env.new(levels: [0,1,0], times:[sample_atk, sample_sus], curve: 'lin'))")
 fx.add("osc = osc*env")
 fx.save()
 
-fx = FxList.new("position", "trimPos", {"position": 0, "sus": 1}, order=2)
+fx = effect_manager.new("position", "trimPos", {"position": 0, "sus": 1}, order=2)
 fx.add("osc = osc * EnvGen.ar(Env(levels: [0,0,1], curve: 'step', times: [sus * position, 0]))")
 fx.save()
 
-fx = FxList.new("ring", "ring_modulation", {"ring":0, "ringl":500, "ringh":1500}, order=0)
+fx = effect_manager.new("ring", "ring_modulation", {"ring":0, "ringl":500, "ringh":1500}, order=0)
 fx.add_var("mod")
 fx.add("mod = ring * SinOsc.ar(Clip.kr(XLine.kr(ringl, ringl + ringh), 20, 20000))")
 fx.add("osc = ring1(osc, mod)")
 fx.save()
 
-fx = FxList.new("shift", "pitchshifter", {"shift":0, "shiftsize": 0.1}, order=1)
+fx = effect_manager.new("shift", "pitchshifter", {"shift":0, "shiftsize": 0.1}, order=1)
 fx.add("osc = PitchShift.ar(osc, shiftsize, shift, 0.02, 0.01)")
 fx.save()
 
-fx = FxList.new("comp", "comp", {"comp": 0, "comp_down": 1, "comp_up": 0.8}, order=2)
+fx = effect_manager.new("comp", "comp", {"comp": 0, "comp_down": 1, "comp_up": 0.8}, order=2)
 fx.add("osc = Compander.ar(osc, osc, thresh: comp, slopeAbove: comp_down, slopeBelow: comp_up, clampTime: 0.01, relaxTime: 0.01, mul: 1)")
 fx.save()
 
-fx = FxList.new("sidechain", "sidechain", {"sidechain": 0, "sidechain_atk": 0.05, "sidechain_rel": 0.1, "thresh":0.006}, order=2)
+fx = effect_manager.new("sidechain", "sidechain", {"sidechain": 0, "sidechain_atk": 0.05, "sidechain_rel": 0.1, "thresh":0.006}, order=2)
 fx.add_var("schain")
 fx.add("schain = In.ar(sidechain,1)")
 fx.add("osc = Compander.ar(osc, schain, thresh: thresh, slopeAbove: 0.1, slopeBelow: 1, clampTime: sidechain_atk, relaxTime: sidechain_rel, mul: 1)")
 fx.save()
 
-fx = FxList.new("lofi", "lofi", {"lofi": 0, "lofiwow": 0.5, "lofiamp":0.5}, order=2)
+fx = effect_manager.new("lofi", "lofi", {"lofi": 0, "lofiwow": 0.5, "lofiamp":0.5}, order=2)
 fx.add_var("minWowRate")
 fx.add_var("wowRate")
 fx.add_var("maxDepth")
@@ -255,7 +256,7 @@ fx.add("osc = MoogFF.ar(osc, LinExp.kr(lofi, 0, 1, 1000, 10000), 0)")
 fx.save()
 
 ### need the miSCellaneous Quark, install in SC
-fx = FxList.new("fold", "wavefold", {"fold": 0, "symetry": 1, "smooth": 0.5}, order=2)
+fx = effect_manager.new("fold", "wavefold", {"fold": 0, "symetry": 1, "smooth": 0.5}, order=2)
 fx.add_var("gain")
 fx.add_var("compensationGain")
 fx.add_var("envFollower")
@@ -267,39 +268,39 @@ fx.add("osc = SmoothFoldS.ar((osc + LinLin.kr(symetry, 0, 1, 1, 0)) * LinLin.kr(
 fx.add("osc = LeakDC.ar(osc*ampgain)")
 fx.save()
 
-fx = FxList.new('low','L_Equalizer', { 'low': 1, 'lowfreq': 80}, order=2)
+fx = effect_manager.new('low','L_Equalizer', { 'low': 1, 'lowfreq': 80}, order=2)
 fx.doc("Low shelf Equalizer")
 fx.add('osc = BLowShelf.ar(osc, freq: lowfreq, db: abs(low).ampdb)')
 fx.save()
 
-fx = FxList.new('mid','M_Equalizer', {'mid': 1, 'midfreq': 1000, 'midq': 1}, order=2)
+fx = effect_manager.new('mid','M_Equalizer', {'mid': 1, 'midfreq': 1000, 'midq': 1}, order=2)
 fx.doc("Middle boost Equalizer")
 fx.add('osc = BPeakEQ.ar(osc, freq: midfreq, rq: midq.reciprocal, db: abs(mid).ampdb)')
 fx.save()
 
-fx = FxList.new('high','H_Equalizer', {'high': 1, 'highfreq': 8000}, order=2)
+fx = effect_manager.new('high','H_Equalizer', {'high': 1, 'highfreq': 8000}, order=2)
 fx.doc("High shelf Equalizer")
 fx.add('osc = BHiShelf.ar(osc, freq: highfreq, db: abs(high).ampdb)')
 fx.save()
 
-fx = FxList.new('phaser', 'phaser', {'phaser': 0, 'phaserdepth': 0.5}, order=2)
+fx = effect_manager.new('phaser', 'phaser', {'phaser': 0, 'phaserdepth': 0.5}, order=2)
 fx.add_var("delayedSignal")
 fx.add("delayedSignal = osc")
 fx.add("for(1, 4, {|i| delayedSignal = AllpassL.ar(delayedSignal, 0.01 * 4.reciprocal, LFPar.kr(LinExp.kr(phaser, 0, 1, 0.275, 16), i + 0.5.rand, LinExp.kr(phaserdepth*4.reciprocal, 0, 1, 0.0005, 0.01 * 0.5), LinExp.kr(phaserdepth*4.reciprocal, 0, 1, 0.0005, 0.01 * 0.5)), 0)})")
 fx.add("osc = osc + delayedSignal")
 fx.save()
 
-fx = FxList.new('ringz','Ringmod', {'ringzfreq': 50, 'ringz': 0.1}, order=2)
+fx = effect_manager.new('ringz','Ringmod', {'ringzfreq': 50, 'ringz': 0.1}, order=2)
 fx.doc("Ringmodulation")
 fx.add('osc = Ringz.ar(osc, freq: ringzfreq, decaytime: ringz, mul: 0.05)')
 fx.save()
 
-fx = FxList.new('resonz','resonz', {'rfreq': 50, 'resonz': 0.1}, order=2)
+fx = effect_manager.new('resonz','resonz', {'rfreq': 50, 'resonz': 0.1}, order=2)
 fx.doc("Resonz")
 fx.add('osc = Resonz.ar(osc, freq: rfreq, bwr: resonz)')
 fx.save()
 
-fx = FxList.new('room2', 'reverb_stereo', {'room2': 0, 'mix2': 0.2, 'damp2':0.8, 'revatk':0, 'revsus':1}, order=2)
+fx = effect_manager.new('room2', 'reverb_stereo', {'room2': 0, 'mix2': 0.2, 'damp2':0.8, 'revatk':0, 'revsus':1}, order=2)
 fx.add_var("dry")
 fx.add("dry = osc")
 fx.add("osc = HPF.ar(osc, 100)")
@@ -309,7 +310,7 @@ fx.add("osc = osc * EnvGen.ar(Env([0,1,0], [revatk,revsus], curve: 'welch'))")
 fx.add("osc = SelectX.ar(mix2, [dry, osc])")
 fx.save()
 
-fx = FxList.new('stut', 'stutterfx', {'t_reset': 0, 'stut': 1, 'stutrate':1, 'stutlen':0.02}, order=2)
+fx = effect_manager.new('stut', 'stutterfx', {'t_reset': 0, 'stut': 1, 'stutrate':1, 'stutlen':0.02}, order=2)
 fx.add_var("dry")
 fx.add_var("reset")
 fx.add_var("wet")
@@ -320,14 +321,14 @@ fx.add("wet = ~stutter.(osc, reset, stutlen)")
 fx.add("osc = SelectX.ar(stut, [dry, wet], wrap:1)")
 fx.save()
 
-#fx = FxList.new('vol','volume', {'vol': 1}, order=2)
+#fx = effect_manager.new('vol','volume', {'vol': 1}, order=2)
 #fx.doc("Volume")
 #fx.add("osc = osc*vol")
 #fx.save()
 
 
 # Fx LOOP
-fx = FxList.new('fx1','fxout', {'fx1': 0, 'lpfx1':22000, 'hpfx1':0}, order=2)
+fx = effect_manager.new('fx1','fxout', {'fx1': 0, 'lpfx1':22000, 'hpfx1':0}, order=2)
 fx.doc("FX1 Bus")
 fx.add_var("fxsig")
 fx.add("fxsig = LPF.ar(osc, lpfx1)")
@@ -335,7 +336,7 @@ fx.add("fxsig = HPF.ar(fxsig, hpfx1)")
 fx.add("Out.ar(2, Mix.ar(fxsig*fx1))")
 fx.save()
 
-fx = FxList.new('fx2','fx2out', {'fx2': 0, 'lpfx2':22000, 'hpfx2':0}, order=2)
+fx = effect_manager.new('fx2','fx2out', {'fx2': 0, 'lpfx2':22000, 'hpfx2':0}, order=2)
 fx.doc("FX2 Bus")
 fx.add_var("fxsig")
 fx.add("fxsig = LPF.ar(osc, lpfx2)")
@@ -343,12 +344,12 @@ fx.add("fxsig = HPF.ar(fxsig, hpfx2)")
 fx.add("Out.ar(3, Mix.ar(fxsig*fx2))")
 fx.save()
 
-fx = FxList.new('output','output', {'output': 0}, order=2)
+fx = effect_manager.new('output','output', {'output': 0}, order=2)
 fx.doc("Output select Bus")
 fx.add("Out.ar(output, osc)")
 fx.save()
 
 ###########
 
-Effect.server.setFx(FxList)
+Effect.server.setFx(effect_manager)
